requirements.txt
asgiref==3.8.1
astroid==3.2.2
dill==0.3.8
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==4.2.13
django-admin-sortable2==2.2.1
django-cache-url==3.4.5
django-extensions==3.2.3
django-tinymce==4.1.0
environs==11.0.0
isort==5.13.2
MarkupSafe==2.1.5
marshmallow==3.21.3
mccabe==0.7.0
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
pylint==3.2.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.11.0
Werkzeug==3.0.3
